--- pages/exercise/exerciseData.json
@@ -1,1 +1,0 @@
-[{"text":"text1","soundUrl":"sound1"},{"text":"text2","soundUrl":"sound2"},{"text":"text3","soundUrl":"sound3"},{"text":"text4","soundUrl":"sound4"}]
